FSPT-1370: move list report functinality into reports routes
Separate out reports (list reports) from collection pages

Note this is split into a number of commits to make sure its tracked as
a `mv` and not replacing 5,000 loc.
